test(server): dedupe the real-binary multiprocess harness into server_process.rs (P10.B2)
